Clean-up of MapsDBAppContext
