+++ runtime/src/main/java/org/corfudb/runtime/collections/graphdb/Node.java
@@ -1,8 +1,6 @@
 package org.corfudb.runtime.collections.graphdb;
 
-import java.util.ArrayList;
 import java.util.HashSet;
-import java.util.List;
 import java.util.Set;
 
 /**
